// src/api/rosenConfig.ts
export type rosenConfig = {
    RSN: string;
    RepoNFT: string;
    RWTId: string;
    minBoxValue: string;
    fee: string;
    guardNFT: string;
    cleanupNFT: string;
    cleanupConfirm: number;
}

export const rosenConfig: rosenConfig = {
    RSN: "a2a6c892c38d508a659caf857dbe29da4343371e597efd42e40f9bc99099a516",
    RepoNFT: "a40b86c663fbbfefa243c9c6ebbc5690fc4e385f15b44c49ba469c91c5af0f48",
    RWTId: "3c6cb596273a737c3e111c31d3ec868b84676b7bad82f9888ad574b44edef267",
    minBoxValue: "1100000",
    fee: "1100000",
    guardNFT: "a6ac381e6fa99929fd1477b3ba9499790a775e91d4c14c5aa86e9a118dfac853",
    cleanupNFT: "ac0c16cee8caefd0814c5568a9d50bea51fec88e8abbd3ac8347b94d054a1b65",
    cleanupConfirm: 0
};